Incomplete - # 22513: missing ValidTime in the GML tag library http://team.ceda.ac.uk/trac/ceda/ticket/22513
Even if this release adds the required tag, the most important difference with the version 0.0.2 is the library itself which has been deeply changed to add more flexibility for future development. 
The main difference is that now the class is not composed by classes per markup type but from createXXX function per markup type which directly return an Element instance.

Another major add is a rich collection of unittests